pyproject.toml
[tool.poetry]
name = "airflow2-dynamic-dags"
version = "0.0.1"
description = "airflow v2 dynamic dags"
authors = ["Savio Teles"]

[tool.poetry.dependencies]
python = ">=3.8.10,<3.9"
apache-airflow = {version = "^2.2.5"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
